pyproject.toml
[project]
name = "workshop-engenharia-de-dados"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.4"
dependencies = [
    "fsspec>=2026.3.0",
    "google-auth>=2.49.2",
    "google-cloud-bigquery>=3.41.0",
    "pandas>=3.0.2",
    "paramiko>=4.0.0",
    "polars>=1.40.0",
    "pyarrow>=24.0.0",
    "pymysql>=1.1.2",
    "sqlalchemy>=2.0.49",
    "xlsxwriter>=3.2.9",
]
